Added "Browse all" button
I wanted to have a Button to display all entries of the index.
Sometimes I just need an inspiration, or a moment to find the right word to look for. That's when I like to browse a list of possible entries.
Since it's only 400 entries, a Pagination function is not really necessary - but that could be a further improvement.